+++ platform/linux/native.h
@@ -9,6 +9,7 @@ typedef struct {
 	GtkWidget *vbox;
 	GtkWidget *menubar;
 	WebKitWebView *view;
+	GtkAccelGroup *accels;
 	char *id;
 	int maximized;
 	int fullscreen;
@@ -37,7 +38,8 @@ void vitra_win_eval(VitraWin *w, const char *js);
 void vitra_win_close(VitraWin *w);
 void vitra_win_free(VitraWin *w);
 void vitra_win_clear_menu(VitraWin *w);
-void vitra_win_add_menu_item(VitraWin *w, const char *menu_label, const char *item_id, const char *item_label);
+void vitra_win_add_menu_item(VitraWin *w, const char *menu_label, const char *item_id, const char *item_label, const char *shortcut);
+int vitra_win_activate_accel(VitraWin *w, const char *shortcut);
 
 char *vitra_clip_get(void);
 void vitra_clip_set(const char *text);